+++ ViewModels/RentalStartVM.cs
@@ -5,7 +5,7 @@
         public string VehicleLicensePlateNumber { get; set; }
         public DateTime RentalStart { get; set; }
         public string CustomerNumber { get; set; }
-        public int StartMileage { get; set; }
+        public float StartMileage { get; set; }
 
     }
 }
